requirements.txt
reportlab==4.2.5
python-docx==1.1.2
Pillow==10.1.0
docx2pdf==0.1.8